pyproject.toml
[project]
name = "pytest-autograder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest>=8.3.4",
    "pyzmq>=26.2.1",
    "six>=1.17.0",
    "tblib>=3.0.0",
]
